backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
duckdb==0.10.3
redis==5.0.4
python-dotenv==1.0.1
groq==1.0.0
python-multipart==0.0.9
pandas==2.2.2
pydantic==2.7.1
